// frontend/src/shared/api/generated/models/client-profile-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Pied Piper Freelance API
 * API для биржи фриланса (MVP)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ClientJobDTO } from './client-job-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { ClientReviewDTO } from './client-review-dto';

export interface ClientProfileDTO {
    'profileId'?: number;
    'accountId'?: number;
    'email'?: string;
    'firstName'?: string;
    'lastName'?: string;
    'fullName'?: string;
    'bio'?: string;
    'avatarUrl'?: string;
    'totalSpent'?: number;
    'postedJobs'?: number;
    'activeJobs'?: number;
    'completedJobs'?: number;
    'totalReviewsGiven'?: number;
    'averageRatingGiven'?: number;
    'memberSince'?: string;
    'lastLogin'?: string;
    'accountStatus'?: string;
    'recentJobs'?: Array<ClientJobDTO>;
    'recentReviews'?: Array<ClientReviewDTO>;
}