// frontend/src/shared/api/generated/models/job-skill-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Pied Piper Freelance API
 * API для биржи фриланса (MVP)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface JobSkillDTO {
    'skillId'?: number;
    'skillName'?: string;
    'skillCategory'?: string;
}